avoid code duplication of DummyInplace
