+++ src/workarounds/fft_generic.jl
@@ -90,17 +90,3 @@ function generic_plan_bfft(data::AbstractArray{T, 3}) where T
                     FourierTransforms.plan_bfft(data[1, :, 1]),
                     FourierTransforms.plan_bfft(data[1, 1, :])], T(1))
 end
-
-
-# A dummy wrapper around an out-of-place FFT plan to make it appear in-place
-# This is needed for some generic FFT implementations, which do not have in-place plans
-struct DummyInplace{opFFT}
-    fft::opFFT
-end
-LinearAlgebra.mul!(Y, p::DummyInplace, X) = (Y .= mul!(similar(X), p.fft, X))
-LinearAlgebra.ldiv!(Y, p::DummyInplace, X) = (Y .= ldiv!(similar(X), p.fft, X))
-
-import Base: *, \, length
-*(p::DummyInplace, X) = p.fft * X
-\(p::DummyInplace, X) = p.fft \ X
-length(p::DummyInplace) = length(p.fft)
